+++ src/main/java/fr/eni/tp/filmotheque/bo/Personne.java
@@ -8,7 +8,7 @@ public abstract class Personne implements Serializable {
 	 * Numéro de sérialisation
 	 */
 	private static final long serialVersionUID = 1L;
-	private long id;
+	private Integer id;
 	private String nom;
 	private String prenom;
 
@@ -20,16 +20,16 @@ public abstract class Personne implements Serializable {
 		this.prenom = prenom;
 	}
 
-	public Personne(long id, String nom, String prenom) {
+	public Personne(Integer id, String nom, String prenom) {
 		this(nom, prenom);// Appel du constructeur précédent
 		this.id = id;
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
